test: LadderBoard 테스트 케이스 작성
