+++ src/test/java/nextstep/ladder/domain/player/TestPlayers.java
@@ -1,0 +1,9 @@
+package nextstep.ladder.domain.player;
+
+public class TestPlayers {
+
+    private TestPlayers() {}
+
+
+
+}
